Use Glimmer component backing class rather than a templateOnly component
When writing documentation it's super handy to be able to write a good
usage example without having to write your own Glimmer component backing
class for the example/template.

Template only components are great for this, but sometimes you would
like to use the `mut` or `set` helper to help to illustrate the
functionality of the component a little better, again without having to
write an additional Glimmer component backing class just to you can
access `this` from within the component.

This commit changes the template backing code to use a full, but empty,
`@glimmer/component` class rather than `templateOnly()`. This means your
usage example can reference `this` and then use either `mut` or `set`
helpers.

The only downside I can see here is that using `templateOnly` is
slightly lighter-weight than using a full glimmer component class, but
in the context of writing documentation, theres something to be said for
making writing documentation as friction free as possible versus making
things light weight.